web: map function name()
dawscript is hardcore on hiding what is the running DAW by presenting a uniform
interface. However Bitwig does not implement a method for queyring parameter
ranges in their Java API, therefore an ugly hack is used that sets param raw
values to 0, then 1 and queries the scaled values. This means changes are
done (and undone) to the loaded project, so this artificial get_parameter_range
function must be used with extra care, probably making exceptions on the client
code. Possible risk of blowing up speakers!